🔨refactor: Refactor main.ts configurations into separate files under the 'setting' folder. Rename s3-service to s3. Rename swagger.ts to api-response.decorator.ts Rename Serializer.ts. to session.serializer.ts Rename GoogleGuard.ts to google-auth.guard.ts
