+++ src/app.service.ts
@@ -4,6 +4,6 @@ import { Injectable } from '@nestjs/common';
 export class AppService {
   
   getHello(): string {
-    return 'CodingSupporter_Backend_0.1';
+    return 'CodingSupporter_Backend_0.9';
   }
 }
