+++ src/auth/auth.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { AuthController } from './auth.controller';
-import { GoogleStrategy } from './utils/GoogleStrategy';
+import { GoogleStrategy } from './google/google.strategy';
 import { AuthService } from './auth.service';
-import { SessionSerializer } from './utils/Serializer';
+import { SessionSerializer } from './session.serializer';
 import { UserModule } from 'src/user/user.module';
 
 @Module({
