+++ src/main.ts
@@ -1,11 +1,11 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { ValidationPipe } from '@nestjs/common';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import { httpsOptions } from './setting/httpsOptions';
 import { setUpSession } from './setting/init.sesstion';
 import { cors } from './setting/cors';
 import * as cookieParser from 'cookie-parser';
+import { setUpSwagger } from './setting/swagger';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule,{
@@ -24,13 +24,7 @@ async function bootstrap() {
     forbidNonWhitelisted: true,
   }));
 
-  const config = new DocumentBuilder()
-  .setTitle('CodingSupporter')
-  .setDescription('CodingSupporterAPI')
-  .setVersion('0.1')
-  .build();
-  const document = SwaggerModule.createDocument(app, config)
-  SwaggerModule.setup('documentation',app,document)
+  setUpSwagger(app)
   
   await app.listen(3000);
 }
