+++ src/s3/s3.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { S3Service } from './s3.service';
+@Module({
+  providers: [S3Service],
+  exports: [S3Service]
+})
+export class S3Module {}
