+++ src/s3/s3.service.spec.ts
@@ -1,15 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { S3ServiceService } from './s3-service.service';
+import { S3Service } from './s3.service';
 
-describe('S3ServiceService', () => {
-  let service: S3ServiceService;
+describe('S3Service', () => {
+  let service: S3Service;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [S3ServiceService],
+      providers: [S3Service],
     }).compile();
 
-    service = module.get<S3ServiceService>(S3ServiceService);
+    service = module.get<S3Service>(S3Service);
   });
 
   it('should be defined', () => {
