+++ src/s3/s3.service.ts
@@ -3,7 +3,7 @@ import { ConfigService } from '@nestjs/config';
 import * as AWS from 'aws-sdk';
 import { v4 as uuid } from 'uuid';
 @Injectable()
-export class S3ServiceService {
+export class S3Service {
   private s3: AWS.S3;
   private bucketName:string;
